SAP
In this example, we establish a connection to SAP using the provided connection parameters. We then call the BAPI_COMPANYCODE_GETLIST function module in SAP to retrieve a list of company codes. The result is returned as a dictionary, from which we extract the COMPANYCODE_LIST data. We convert this data to a Pandas DataFrame, and finally, we display the first few rows of the DataFrame using head().